fix: rename cast.py to cast_manager.py to avoid HA cast platform discovery
HA's cast integration auto-discovers any cast.py file in a custom
integration and calls _register_cast_platform() on it, expecting a
proper cast platform interface. Our internal cast session manager code
doesn't implement that interface, causing 'Invalid cast platform'
errors on every startup and reload.

Renaming to cast_manager.py prevents the discovery entirely.
Also update VerifyEntity in copilot instructions to correct sensor name.
